Refactor auth configuration
